requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# AI & LLM
anthropic>=0.78.0
openai>=2.17.0

# Utilities
numpy>=2.4.0
pydantic>=2.12.0
pydantic-settings>=2.12.0

# Development
pytest>=9.0.0
black>=26.1.0

# Optional: Simulation (맥북에서 빌드 오류 발생 가능, 나중에 설치)
# pybullet>=3.2.6

# Optional: Audio (음성 기능 구현 시 필요)
# pyaudio>=0.2.14
# sounddevice>=0.4.6
